backend/requirements.txt
# 核心框架
fastapi==0.103.1
uvicorn==0.23.2
celery==5.3.6

# 数据库工具
sqlalchemy==2.0.21
python-dotenv==1.0.0

# 数据处理
pandas==2.1.1
numpy==1.26.0
openpyxl==3.1.2         # 处理 Excel 文件（若仅用 CSV 可移除）

# 其他工具
requests==2.31.0
python-multipart==0.0.6  # FastAPI 处理表单数据/文件上传
